feat: 支持win系统软件搜索
