+++ package.json
@@ -11,7 +11,7 @@
     "dev": "node .electron-vue/dev-runner.js",
     "rebuild": " ./node_modules/.bin/electron-rebuild",
     "rebuild_win": "npm rebuild --runtime=electron --target=11.4.10 --disturl=https://atom.io/download/atom-shell --abi=85",
-    "rebuild_linux": "npm rebuild --runtime=electron --target=12.0.15 --disturl=https://atom.io/download/atom-shell --abi=87"
+    "rebuild_linux": "npm rebuild --runtime=electron --target=11.4.10 --disturl=https://atom.io/download/atom-shell --abi=85"
   },
   "build": {
     "asar": true,
@@ -56,7 +56,9 @@
     },
     "linux": {
       "icon": "build/icons/",
-      "publish": ["github"]
+      "publish": [
+        "github"
+      ]
     }
   },
   "dependencies": {
@@ -67,6 +69,7 @@
     "download-git-repo": "^3.0.2",
     "electron-is-dev": "^2.0.0",
     "electron-store": "^8.0.0",
+    "icon-extractor": "^1.0.3",
     "iohook": "^0.9.3",
     "is-chinese": "^1.4.2",
     "jian-pinyin": "^0.2.3",
@@ -109,7 +112,7 @@
     "css-loader": "^0.28.11",
     "del": "^3.0.0",
     "devtron": "^1.4.0",
-    "electron": "^12.0.15",
+    "electron": "^11.4.10",
     "electron-builder": "22.10.5",
     "electron-debug": "^1.5.0",
     "electron-devtools-installer": "^2.2.4",
@@ -140,7 +143,7 @@
   "iohook": {
     "targets": [
       "node-83",
-      "electron-87"
+      "electron-85"
     ],
     "platforms": [
       "darwin",
